refactor classification of workers to own system
